backend/requirements.txt
Flask
flask-cors
gunicorn
python-dotenv
requests
serpapi
google-generativeai
langchain
langchain-google-genai
pydantic>=2
reportlab
sentence-transformers
faiss-cpu
pymupdf
python-docx
Pillow
numpy
